ref(onboarding): Split ruby onboarding doc into multiple files
