+++ static/app/gettingStartedDocs/ruby/ruby/index.tsx
@@ -1,0 +1,17 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {CrashReportWebApiOnboarding} from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
+
+import {logs} from './logs';
+import {onboarding} from './onboarding';
+import {profiling} from './profiling';
+
+const docs: Docs = {
+  onboarding,
+  crashReportOnboarding: CrashReportWebApiOnboarding,
+  profilingOnboarding: profiling(),
+  logsOnboarding: logs({
+    docsPlatform: 'ruby',
+  }),
+};
+
+export default docs;
