+++ static/app/gettingStartedDocs/ruby/ruby/onboarding.spec.tsx
@@ -4,7 +4,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './ruby';
+import docs from '.';
 
 describe('getting started with ruby', () => {
   it('renders errors onboarding docs correctly', () => {
